# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Um repositorio de autores e suas obras",
        "description": "Um repositorio para agrupar todos os autores e suas obras conhecidas da humanidade",
        "termsOfService": "https://minhaapi.com/terms",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/"
        }
    ],
    "paths": {
...
        "/User": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Criar um usuario",
                "description": "Cria um usuario",
                "operationId": "7b6848ef7fc0da35b0928f26d759979b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "administrador": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "nome do usuario",
                                            "format": "string",
                                            "default": "daniinho"
                                        },
                                        "email": {
                                            "description": "email do usuario",
                                            "format": "string",
                                            "default": "deyllonp986@gmail.com"
                                        },
                                        "password": {
                                            "description": "senha do usuario",
                                            "format": "string",
                                            "default": "amanha27"
                                        },
                                        "administrador": {
                                            "description": "se um usuario é administrador ou não",
                                            "format": "boolean",
                                            "default": "S"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/User/{User}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Atualizar um Usuario",
                "description": "Atualizar um Usuario",
                "operationId": "a72022dacef381050695c369598355f0",
                "parameters": [
                    {
                        "name": "User",
                        "in": "path",
                        "description": "Id do usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "administrador": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "email": {
                                            "description": "email do usuario",
                                            "format": "string",
                                            "default": "danilp864@gmail.com"
                                        },
                                        "name": {
                                            "description": "nome do usuario",
                                            "format": "string",
                                            "default": "dabip"
                                        },
                                        "administrador": {
                                            "description": "se o usuario é administrador ou não",
                                            "format": "boolean",
                                            "default": "N"
                                        },
                                        "password": {
                                            "description": "senha do usuario",
                                            "format": "string",
                                            "default": "amazingk"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Pegue o token indo na operação login e passando o email e senha",
                "name": "Token JWT",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}